Change the style of the selection checkboxes in the Scoreboard
They're now just table cells which are "simulating" checkboxes through CSS.
This made them more flexible and easy to style, allowing us to set a background
color and to make them larger.

The introduction of different colors for selection hasn't been reflected yet in
the Overview (which still shows all selected users in blue). This is because I'm
planning to rewrite the Overview (almost) from scratch.