Improve how RWS handles connection losses
Properly support the Last-Event-ID HTTP header (and URL argument) in
Server-Sent-Events reqeuests to avoid the client missing some messages.
This requires the server to keep a buffer of recently sent events.

Also show a "network status indicator" showing the status of the SSE
connection and allowing the user to manually retry the connection if the
browser failed (or reloading the page if the local data is outdated).